Create README.md
By clustering different descriptions and locations together, a location and content-based recommendation system can be created using only K-means algorithm and text vectorization. Ideally, this system would grab a user’s latitude and longitude location and give a restaurant recommendation based on previous restaurant that the user liked in that area. Some potential improvements to the recommendation system is being able to filter out repeats that have the same restaurant name but with a different way of titling the restaurant name. For example, Starbucks is sometimes titled as Starbucks Co. and other times it is titled as Starbucks Coffee. Another potential improvement would be to use an alternative to K-means clustering for the text-based recommendation because, for example, it may be better to use cosine similarity rather than K-means clustering to determine whether a restaurant should be grouped with each other. Additionally, another potential improvement would be to implement a top scored based recommendation instead of the randomly generated recommendations made by the K-means clustering algorithm.